+++ src/main/java/smartcity/stations/StationStrategy.java
@@ -24,7 +24,7 @@ public class StationStrategy {
     final private Map<String, PedestrianArrivalInfo> pedestrianOnStationMap = new HashMap<>();
 
     //private final OSMStation stationOSMNode;
-    public StationStrategy(OSMStation stationOSMNode, long agentId) {
+    public StationStrategy(OSMStation stationOSMNode, int agentId) {
         MainContainerAgent.osmStationIdToStationNode.put(stationOSMNode.getId(), new StationNode(stationOSMNode.getLat(),
                 stationOSMNode.getLon(), Long.toString(stationOSMNode.getId()), agentId));
     }
